refactor: Move PlayerFinder to player package
